FX2 fw: fix wait times to be actually in line with real time.
git-svn-id: svn://svn.code.sf.net/p/jaer/code@5250 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ devices/firmware/CypressFX2/SeeBetterLogic_DVS128/xsvf_player/ports.h
@@ -32,7 +32,7 @@ extern void pulseClock();
 /* read the next byte of data from the xsvf file */
 extern void readByte(unsigned char *ucdata);
 
-extern void waitTime(int microsec);
+extern void waitTime(long microsec);
 
 extern void resetDataArray(unsigned char *newDataArray);
